+++ love/src/components/CSCSummary/CSCExpanded/CSCExpanded.jsx
@@ -152,6 +152,27 @@ export default class CSCExpanded extends PureComponent {
     },
   };
 
+  static bigCameraOfflineDetailedStates = {
+    0: {
+      name: 'UNKNOWN',
+      userReadable: 'Unknown',
+      char: 'U',
+      class: styles.unknown,
+    },
+    1: {
+      name: 'AVAILABLE',
+      userReadable: 'Offline Available',
+      char: 'A',
+      class: styles.ok,
+    },
+    2: {
+      name: 'PUBLISH_ONLY',
+      userReadable: 'Offline Publish Only',
+      char: 'P',
+      class: styles.warning,
+    },
+  };
+
   static validLogLevel = {
     DEBUG: 10,
     INFO: 20,
@@ -169,11 +190,21 @@ export default class CSCExpanded extends PureComponent {
   };
 
   static validState = {
+    enterControl: 'OFFLINE',
     start: 'STANDBY',
     enable: 'DISABLED',
     disable: 'ENABLED',
     standby: 'DISABLED or FAULT',
+    exitControl: 'STANDBY',
   };
+
+  setListOfStateTransitionCommands(name) {
+    let commands = ['start', 'enable', 'disable', 'standby', 'exitControl'];
+    if (name.search(/[A-Z]{2}Camera/) != -1) {
+      commands.unshift('enterControl');
+    }
+    return commands;
+  }
 
   setSummaryStateCommand(option) {
     const { configurationOverride } = this.state;
@@ -225,6 +256,7 @@ export default class CSCExpanded extends PureComponent {
       name,
       salindex,
       summaryStateData,
+      offlineDetailedStateData,
       softwareVersions,
       configurationsAvailable: configurationsAvailableData,
       configurationApplied: configurationAppliedData,
@@ -287,7 +319,13 @@ export default class CSCExpanded extends PureComponent {
       heartbeatTitle += `Last seen: ${timeDiffText}`;
     }
 
+    const isBigCamera = name.search(/[A-Z]{2}Camera/) != -1;
+    const offlineDetailedStateValue = offlineDetailedStateData ? offlineDetailedStateData.substate.value : 0;
+    const isOffline = summaryState.name === 'OFFLINE';
+    const bigCameraOfflineDetailedState = CSCExpanded.bigCameraOfflineDetailedStates[offlineDetailedStateValue];
+
     const isSimulated = simulationMode?.mode.value > 0;
+
     return (
       <div className={styles.CSCGroupContainer}>
         <div className={styles.CSCExpandedContainer}>
@@ -322,6 +360,16 @@ export default class CSCExpanded extends PureComponent {
                       {summaryState.name}
                     </span>
                   </div>
+                  {isBigCamera && isOffline && (
+                    <div>
+                      <span
+                        className={[styles.summaryState, bigCameraOfflineDetailedState.class].join(' ')}
+                        title={bigCameraOfflineDetailedState.userReadable}
+                      >
+                        {bigCameraOfflineDetailedState.name}
+                      </span>
+                    </div>
+                  )}
                   <div className={styles.heartbeatIconWrapper}>
                     <HeartbeatIcon className={styles.heartbeatIcon} status={heartbeatStatus} title={heartbeatTitle} />
                   </div>
@@ -362,7 +410,7 @@ export default class CSCExpanded extends PureComponent {
                 <div className={styles.breadcrumContainer}>
                   <div className={styles.titlePadding}>Select State transition Command:</div>
                   <Select
-                    options={['start', 'enable', 'disable', 'standby']}
+                    options={this.setListOfStateTransitionCommands(name)}
                     onChange={(option) => this.setSummaryStateCommand(option.value)}
                     value=""
                     placeholder="Select state"