+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Version History
 v6.7.3
 ------
 
+* Add exitControl for all CSCs and enterControl to XXCamera CSCs in expanded view. `<https://github.com/lsst-ts/LOVE-frontend/pull/711>`_
+* Add Offline detailed stated for big cameras in CSC detailed and expanded views. `<https://github.com/lsst-ts/LOVE-frontend/pull/711>`_
 * Fix MTDomeShutter component interpreting MTDome_apertureShutter value erroneously. `<https://github.com/lsst-ts/LOVE-frontend/pull/710>`_
 * Roll back critical alarm loop for Watcher alarm sounds. `<https://github.com/lsst-ts/LOVE-frontend/pull/709>`_
 * Fix HVAC Facility Map display after XML 23 changes. `<https://github.com/lsst-ts/LOVE-frontend/pull/707>`_